+++ docs/pages/examples/dashboard/index.tsx
@@ -4,7 +4,6 @@ import {
   Grid,
   HStack,
   Heading,
-  Icon,
   IconButton,
   Menu,
   MenuButton,
@@ -15,7 +14,7 @@ import {
   VStack,
 } from "@yamada-ui/react"
 import type { SegmentedControlItem, StackProps } from "@yamada-ui/react"
-import { ChevronDown } from "lucide-react"
+import { ChevronDown } from "@yamada-ui/lucide"
 import { memo } from "react"
 import type { FC } from "react"
 import { Cards } from "./cards"
@@ -84,7 +83,7 @@ const Content: FC<ContentProps> = memo(({ children, ...rest }) => {
               variant="ghost"
               display={{ base: "none", md: "flex" }}
             >
-              <Icon as={ChevronDown} fontSize="2xl" />
+              <ChevronDown fontSize="2xl" />
             </MenuButton>
 
             <MenuList>