+++ packages/components/tooltip/tests/tooltip.test.tsx
@@ -1,7 +1,7 @@
 import { a11y, fireEvent, render, screen, waitFor } from "@yamada-ui/test"
 import { Tooltip } from "../src/tooltip"
 
-describe("<Tooltip/>", () => {
+describe("<Tooltip />", () => {
   test("should pass a11y test", async () => {
     await a11y(
       <Tooltip label="Tooltip hover">
@@ -76,5 +76,25 @@ describe("<Tooltip/>", () => {
 
     const tooltip = screen.queryByRole("tooltip")
     expect(tooltip).toBeNull()
+  })
+
+  test("should not render label text when `Escape` pressed", async () => {
+    render(
+      <Tooltip label="Tooltip hover">
+        <span>Hover</span>
+      </Tooltip>,
+    )
+
+    fireEvent.pointerEnter(screen.getByText("Hover"))
+
+    await waitFor(() => {
+      expect(screen.getAllByText("Tooltip hover")[1]).toBeInTheDocument()
+    })
+
+    fireEvent.keyDown(screen.getByText("Hover"), { key: "Escape" })
+
+    await waitFor(() => {
+      expect(screen.queryByRole("tooltip")).toBeNull()
+    })
   })
 })