+++ stories/system/styles/breakpoint.stories.tsx
@@ -1,4 +1,12 @@
-import { Box, useBreakpoint, useBreakpointValue } from "@yamada-ui/react"
+import {
+  Box,
+  extendConfig,
+  Text,
+  UIProvider,
+  useBreakpoint,
+  useBreakpointValue,
+} from "@yamada-ui/react"
+import { useRef, type FC } from "react"
 
 export default {
   title: "System / Styles / Breakpoint",
@@ -47,6 +55,66 @@ export const useHook = () => {
       transitionDuration="slower"
     >
       The current breakpoint is "{breakpoint}"
+    </Box>
+  )
+}
+
+export const withContainer = () => {
+  const containerRef = useRef<HTMLDivElement>(null)
+  const breakpoint = useBreakpoint()
+
+  const config = extendConfig({
+    breakpoint: {
+      containerRef,
+      identifier: "@container",
+    },
+  })
+
+  const App: FC = () => {
+    const breakpoint = useBreakpoint()
+
+    return (
+      <Text
+        color={{
+          base: "red.500",
+          "2xl": "cyan.500",
+          xl: "blue.500",
+          lg: "green.500",
+          md: "yellow.500",
+          sm: "purple.500",
+        }}
+      >
+        The container breakpoint is "{breakpoint}"
+      </Text>
+    )
+  }
+
+  return (
+    <Box
+      ref={containerRef}
+      containerType="inline-size"
+      resize="inline"
+      overflow="auto"
+      w="full"
+      p="md"
+      borderWidth="1px"
+      rounded="md"
+    >
+      <Text
+        color={{
+          base: "red.500",
+          xl: "blue.500",
+          lg: "green.500",
+          md: "yellow.500",
+          sm: "purple.500",
+        }}
+      >
+        The root breakpoint is "{breakpoint}"
+      </Text>
+
+      <UIProvider config={config}>
+        <App />
+      </UIProvider>
     </Box>
   )
 }