+++ packages/components/switch/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @yamada-ui/switch
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies [[`98bf410`](https://github.com/yamada-ui/yamada-ui/commit/98bf410695434f9c76da4e9f9635dcbfde5b9856), [`d772f23`](https://github.com/yamada-ui/yamada-ui/commit/d772f2339b5ff6e5bfc11fb979566cd38a04df37)]:
+  - @yamada-ui/core@1.9.0
+  - @yamada-ui/checkbox@1.1.0
+  - @yamada-ui/motion@2.0.1
+
+## 1.1.0
+
+### Minor Changes
+
+- [#2121](https://github.com/yamada-ui/yamada-ui/pull/2121) [`ad2a242`](https://github.com/yamada-ui/yamada-ui/commit/ad2a242de2fdd5a4371a7069604a89e0e3cf6f7c) Thanks [@koralle](https://github.com/koralle)! - Enhance `<Switch >` component accessibility
+
+### Patch Changes
+
+- Updated dependencies [[`76d5a75`](https://github.com/yamada-ui/yamada-ui/commit/76d5a7580de91d9226ea1a41286abc0999278b21), [`f10ae87`](https://github.com/yamada-ui/yamada-ui/commit/f10ae8763e87a5c45181d9c99c173c399a048a3b), [`2a30dae`](https://github.com/yamada-ui/yamada-ui/commit/2a30dae39c2bb257b5f5a3c7f077612388bb972b), [`851cb33`](https://github.com/yamada-ui/yamada-ui/commit/851cb3313f939b3b2e85edddab60f79bade2e981), [`28a276f`](https://github.com/yamada-ui/yamada-ui/commit/28a276fc7c11bce28424e763a0f5fba8b5fd8fd0), [`d4833c9`](https://github.com/yamada-ui/yamada-ui/commit/d4833c990016fcb6330ed343b9b464272a1e1053)]:
+  - @yamada-ui/utils@1.3.0
+  - @yamada-ui/core@1.8.0
+  - @yamada-ui/motion@2.0.0
+  - @yamada-ui/checkbox@1.0.35
+
 ## 1.0.34
 
 ### Patch Changes