+++ packages/hooks/use-media-query/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @yamada-ui/use-media-query
 
+## 1.0.38
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @yamada-ui/providers@1.1.5
+
+## 1.0.37
+
+### Patch Changes
+
+- Updated dependencies [[`76d5a75`](https://github.com/yamada-ui/yamada-ui/commit/76d5a7580de91d9226ea1a41286abc0999278b21), [`2a30dae`](https://github.com/yamada-ui/yamada-ui/commit/2a30dae39c2bb257b5f5a3c7f077612388bb972b), [`d4833c9`](https://github.com/yamada-ui/yamada-ui/commit/d4833c990016fcb6330ed343b9b464272a1e1053)]:
+  - @yamada-ui/utils@1.3.0
+  - @yamada-ui/providers@1.1.4
+
 ## 1.0.36
 
 ### Patch Changes